requirements.txt
africastalking==1.2.4
asgiref==3.4.1
autopep8==1.5.7
backcall==0.2.0
beautifulsoup4==4.9.3
bootstrap4==0.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cloudinary==1.26.0
colorama==0.4.4
config==0.5.0.post0
contextlib2==21.6.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.0
django-allauth==0.45.0
django-ckeditor==6.1.0
django-clean==0.1.2
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-daraja==1.2.0
django-elasticsearch-dsl==7.2.0
django-environ==0.4.5
django-extensions==3.1.3
django-heroku==0.3.1
django-js-asset==1.2.2
django-phonenumber-field==5.2.0
django-qr-code==2.2.0
django-ranged-response==0.2.0
django-rest-framework==0.1.0
django-rq==2.4.1
django-simple-captcha==0.5.14
django-taggit==1.5.1
django-user-visit==0.4.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.2
drf-yasg==1.20.0
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
ipdb==0.13.9
ipython==7.25.0
ipython-genutils==0.2.0
isort==5.9.3
itypes==1.2.0
jaraco.context==4.0.0
jedi==0.18.0
Jinja2==3.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
more-itertools==8.8.0
mysqlclient==2.0.3
ndg-httpsclient==0.5.1
numpy==1.21.1
oauth2-provider==0.0
oauthlib==3.1.1
packaging==21.0
panda==0.3.1
pandas==1.3.1
parso==0.8.2
phonenumbers==8.12.28
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.19
psycopg2==2.8.6
pusher==3.0.0
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
PyMySQL==1.0.2
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==0.18.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2021.1
qrcode==7.3
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rq==1.9.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
schema==0.7.4
segno==1.3.3
simplejson==3.17.3
simplejwt==2.0.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
twilio==6.62.1
typing==3.7.4.3
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.6
user-agents==2.2.0
users==1.0.dev0
wcwidth==0.2.5
whitenoise==5.3.0
wikipedia==1.4.0
wolframalpha==5.0.0
xmltodict==0.12.0
